public/GCP_codefiles/gcp_nxs-function/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-aiplatform
vertexai
pandas
numpy
scikit-learn
google-cloud-logging
